Perform a HEAD before the actual request. Close #340

+++ src/FeedIo/Adapter/NullResponse.php
@@ -11,7 +11,7 @@ class NullResponse implements ResponseInterface
     /**
      * @return int
      */
-    public function getDuration(): int
+    public function getDuration(): float
     {
         return 0;
     }
